Fixing of the design and style systems
I added and edited the Login page so it is similar to the add foods page, meaning that there is a box for both username and password to be entered.  I also added a new color gradient to it so that it stands out from the other pages.  I also added images to represent each of the social media platforms, as well as adding a link directly to the Facebook page.  I also added a logo for the restaurant itself at the bottom of each page next to the place filler contact information.